Move CmInstanceTypeGen to component_model.rs
